backend/requirements.txt
fastapi
uvicorn[standard]
python-multipart
google-genai
python-dotenv
python-docx
python-pptx
openpyxl
py7zr
pefile
olefile
xlrd
httpx
elevenlabs
lancedb
pandas
